+++ src/DG_OperatorMatrices_2d/MassMatrix.h
@@ -10,26 +10,26 @@ using namespace std;
 #ifndef MassMatrix_HPP
 #define MassMatrix_HPP
 
-void massMatrix(double *MassMatrix,unsigned N)
+void massMatrix(float *MassMatrix,unsigned N)
 {
-    double Poly[N+1][N+1];
-    double **poly;
-    poly   =   new double*[N+1];
+    float Poly[N+1][N+1];
+    float **poly;
+    poly   =   new float*[N+1];
     lagrangePolynomials(*Poly,N);
     unsigned i,j;
 
     for(i=0;i<=N;i++)
     {
-        poly[i] =   new double[N+1];
-        memcpy(poly[i],Poly[i],(N+1)*sizeof(double));
+        poly[i] =   new float[N+1];
+        memcpy(poly[i],Poly[i],(N+1)*sizeof(float));
     }
 
-    function<double(double)> eval;
+    function<float(float)> eval;
     for(i=0;i<=N;i++)
     {
         for(j=0;j<=N;j++)
         {
-            eval = [&poly,&i,&j,&N](double x){return (((polyEval(poly[i],N,x))*(polyEval(poly[j],N,x))));};
+            eval = [&poly,&i,&j,&N](float x){return (((polyEval(poly[i],N,x))*(polyEval(poly[j],N,x))));};
             MassMatrix[i*(N+1)+j] = lobattoIntegration(-1.0,1.0,N+1,eval);
         }
     }
@@ -41,9 +41,9 @@ void massMatrix(double *MassMatrix,unsigned N)
     return ;
 }
 
-void twoDMassMatrix(double *MassMatrix, unsigned N)
+void twoDMassMatrix(float *MassMatrix, unsigned N)
 {
-    double m[N+1][N+1];
+    float m[N+1][N+1];
     massMatrix(*m,N);
     unsigned i1,i2,j1,j2;
